split rust lib and comparison code
