+++ comparison/Cargo.toml
@@ -1,10 +1,12 @@
 [package]
-name = "fish_hash"
+name = "comparison"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-blake3 = "1.3.1"
-tiny-keccak = { version = "2.0.2", features = ["keccak"] }
+fish_hash = { path = "../rust" }
+
+[build-dependencies]
+cmake = "0.1.50"
