+++ comparison/src/fish_hash_bindings.rs
